+++ NEWS.md
@@ -1,3 +1,5 @@
+# arima2 (development version)
+
 # arima2 3.0.1
 
 * Added arXiv paper to package description