+++ NEWS.md
@@ -1,5 +1,7 @@
 # arima2 (development version)
 
+* Added `...` argument to `aicTable` function. 
+
 # arima2 3.0.1
 
 * Added arXiv paper to package description